requirements.txt
telethon
openai
Pillow
python-dotenv
